Move `BaseballCard` to `lib` module
