+++ android/src/androidTest/java/bbct/android/common/activity/test/BaseballCardDetailsAddCardsTest.java
@@ -28,7 +28,7 @@ import android.widget.Button;
 import android.widget.EditText;
 import bbct.android.common.activity.BaseballCardDetails;
 import bbct.android.common.activity.FragmentTestActivity;
-import bbct.android.common.data.BaseballCard;
+import bbct.data.BaseballCard;
 import bbct.android.common.test.BBCTTestUtil;
 import bbct.android.common.test.BaseballCardCsvFileReader;
 import bbct.android.common.test.DatabaseUtil;
