+++ android/src/androidTest/java/bbct/android/common/activity/test/BaseballCardDetailsPartialInputTest.java
@@ -26,7 +26,7 @@ import android.widget.EditText;
 import bbct.android.common.R;
 import bbct.android.common.activity.BaseballCardDetails;
 import bbct.android.common.activity.FragmentTestActivity;
-import bbct.android.common.data.BaseballCard;
+import bbct.data.BaseballCard;
 import bbct.android.common.test.BBCTTestUtil;
 import bbct.android.common.test.BaseballCardCsvFileReader;
 import java.io.IOException;
