+++ android/src/androidTest/java/bbct/android/common/activity/test/BaseballCardListWithoutDataTest.java
@@ -27,7 +27,7 @@ import android.widget.ListView;
 import bbct.android.common.R;
 import bbct.android.common.activity.BaseballCardList;
 import bbct.android.common.activity.MainActivity;
-import bbct.android.common.data.BaseballCard;
+import bbct.data.BaseballCard;
 import bbct.android.common.test.BBCTTestUtil;
 import bbct.android.common.test.BaseballCardCsvFileReader;
 import bbct.android.common.test.DatabaseUtil;
