+++ android/src/androidTest/java/bbct/android/common/activity/test/FilterCardsCombinationTest.java
@@ -25,7 +25,7 @@ import bbct.android.common.R;
 import bbct.android.common.activity.FilterCards;
 import bbct.android.common.activity.FragmentTags;
 import bbct.android.common.activity.MainActivity;
-import bbct.android.common.data.BaseballCard;
+import bbct.data.BaseballCard;
 import bbct.android.common.test.BBCTTestUtil;
 import bbct.android.common.test.BaseballCardCsvFileReader;
 import bbct.android.common.test.DatabaseUtil;
