+++ android/src/main/java/bbct/android/common/activity/BaseballCardDetails.java
@@ -44,7 +44,7 @@ import android.widget.Spinner;
 import android.widget.Toast;
 import bbct.android.common.R;
 import bbct.android.common.activity.util.DialogUtil;
-import bbct.android.common.data.BaseballCard;
+import bbct.data.BaseballCard;
 import bbct.android.common.provider.BaseballCardContract;
 import bbct.android.common.provider.SingleColumnCursorAdapter;
 import butterknife.ButterKnife;
