+++ android/src/main/java/bbct/android/common/provider/BaseballCardAdapter.java
@@ -33,7 +33,7 @@ import android.widget.SimpleCursorAdapter;
 import bbct.android.common.R;
 import bbct.android.common.activity.BaseballCardList;
 import bbct.android.common.activity.util.BaseballCardMultiChoiceModeListener;
-import bbct.android.common.data.BaseballCard;
+import bbct.data.BaseballCard;
 import bbct.android.common.view.BaseballCardView;
 import butterknife.ButterKnife;
