+++ android/src/main/java/bbct/android/common/provider/BaseballCardContract.java
@@ -22,8 +22,8 @@ import android.content.ContentResolver;
 import android.content.ContentValues;
 import android.database.Cursor;
 import android.net.Uri;
-import android.util.Log;
-import bbct.android.common.data.BaseballCard;
+
+import bbct.data.BaseballCard;
 import java.util.ArrayList;
 import java.util.List;
