+++ lib/src/main/java/bbct/data/BaseballCard.java
@@ -16,7 +16,7 @@
  * You should have received a copy of the GNU General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-package bbct.android.common.data;
+package bbct.data;
 
 import java.io.Serializable;
